+++ CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ## Unreleased
 
+## v2.15.0 (2024-12-30)
+
+### Added
+
+- `<C-d>` and `<C-u>` to go up/down 1/2 page in both normal mode and visual mode.
+
+### Changed
+
+- Avoid displaying big text in preview area by adding PreviewType::TooBigText (over 1MB).
+
+### Fixed
+
+- Rewrite test_has_write_permission by set_readonly.
+
+## v2.14.0 (2024-09-29)
+
+### Changed
+
+- Update dependencies for Rust 1.81.0. MSRV is the same as before (1.74.1).
+
 ## v2.13.0 (2024-04-07)
 
 ### Added